requirements.txt
blis==0.7.3
cachetools==4.2.1
catalogue==2.0.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
cymem==2.0.4
dialogflow==1.1.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.0.0/en_core_web_lg-3.0.0-py3-none-any.whl
Flask==1.1.2
gevent==21.1.2
google-api-core==1.26.0
google-auth==1.27.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.35.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lemminflect==0.2.2
MarkupSafe==1.1.1
murmurhash==1.0.4
nameparser==1.0.6
numpy==1.19.4
packaging==20.9
pathy==0.4.0
plac==1.1.3
pluralizer==1.1.0
preshed==3.0.4
protobuf==3.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.0
rsa==4.7.2
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-legacy==3.0.2
srsly==2.4.1
thinc==8.0.2
tqdm==4.54.0
typer==0.3.2
urllib3==1.26.2
waitress==1.4.4
wasabi==0.8.2
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
